A Simple Reverse Program
A simple program to print a text backwards. This is a must-know for every beginner in Python.
Not only know it, but understand every detail.
This one actually took me a long time to understand, and when i first made it ( i always try to do it my way first) it came up a little different. Still it worked. This is a simpler version.
This exercise covers the 'input()' function, the 'while' statement, the 'len()' function and the ultimate important indexing of strings.